curtin vmtest: skip all skip-by-date when testing packages.
Curtin vmtest skips some tests by date.  This mechanism allows us
to ignore (and not test) known issues, and be reminded of these
skips at a later date.  That works well for trunk.
There, when a failure pops up we can easily fix it.
When when we test packages, though, that is  more difficult.

The change here is to skip all 'skip-by-date' when running
tests of packages.