Switch to using the snapcraft snap over the deb
The two sources for snapcraft differ on behaviour and release
management. We're having trouble aligning our CI with "official" builds
on Launchpad. We've been advised to try the edge snap which has some
non-determinism removed. And we're told that Launchpad is expected to
switch to the snap eventually. So tweak our CI accordingly.