yaml: update job-templates to jobs
Throughout the code we were using job-template instead of job. I am
unsure as to why this occured, but these are uncessary in all, but one
of the cases I found. The result of this was needing to double bracket
quote all variables instead of single quoting them. This makes the
pipeline jobs look rather complex and complicated.