simplestreams: Move to using git and run tox in a container.
Simplestreams has moved from bzr to git, and bzr will stop working
shortly. [1]

It also changes the tests to run in an lxc (bionic) container via
utilization of 'ctool' which was added today to uss-tableflip.

Hopefully running in a container will fix the errors we've been
seeing lately around gpg:
  gpg: can't connect to the agent: IPC connect call failed
  gpg: error getting the KEK: No agent running

[1] https://code.launchpad.net/~smoser/simplestreams/trunk.moved-to-git/+merge/348547