torkoal: set TMPDIR in cloud-init nocloud-kvm tests to use nvme drive.
The jenkins slave 'torkoal' has a tmpdir set up on top of an nvme disk at
/var/lib/jenkins/tmp.  This change will make the nocloud tests use
that for their TMPDIR.  The hope is that this will fix some timeouts
we were seeing in nocloud-kvm tests where the system simply was taking
longer than allowed to boot.

Note that TMPDIR is documented (and tested locally) as passing
through tox without a 'passenv'.