cloud-init: Match test harness version and daily deb version.
When running integration tests, we should match the version of the
test-harness with the version of the deb we're running.  Otherwise
we will get skew and have test failures when a test makes an assertion
that is not availble in the installed package or vice-versa.

The fix here pulls the hash out of the daily deb file name.
   cloud-init_18.4-1964-g9f88125-0ubuntu1+1635~trunk~ubuntu18.04.1_all.deb

It grabs the -g9f88125 there and then checks out cloud-init at gf88125.