curtin: Add a vmtest_filters parameter for proposed testing default to ubuntu only (#17)
* curtin: only test ubuntu releases on -proposed pocket testing

Use --filter to restrict the set of vmtests to ubuntu-only.
Remove some leading whitespace on empty lines.

* curtin: Add a vmtest_filter parameter for proposed testing

We want to only test ubuntu for our proposed testing.  Add a new
parameter to the proposed test and set the default value to
target_distro=ubuntu.  This allows users to customize the job
later so add additional (or remove) the filter as well.

* Drop unneeded comment

* Don't use curly braces for shell local variables