copr: retry failure-prone commands
There is a conflict between running a job made to spot problems and
implementing a retry mechanism that hides them. The decision here is to
retry commands that rely on external resources we do not control (in
this case, CentOS mirrors and build infrastructure), and which are
causing enough false positives to make spotting a real problem actually
harder.