docker-image-size: improve script to avoid leaving stale containers
Changes:
 - trap EXIT and call a container cleanup function
 - use printf(1) instead of echo to print strings with backslash escapes
 - use apt-get instead of apt, following the advice given by apt(8)
 - cosmetic changes when running with `set -x`

While this version of the script is unlikely to leave stale containers,
as the `cleanup` function is normally called twice (the first time
explicitly, the second time by the trap), it will still exit nonzero on
"ZFS dataset is busy" issues. We still don't want to work around them,
as a proper fix is needed.