Replace the curtin-vmtest-sync-images job with a builder (#81)
Replace the curtin-vmtest-sync-images job with a builder

The builder is run before each vmtest run, checks if the images have been
updated recently enough and runs vmtest-sync-images if necessary. This
will update the images on any node the vmtests are run on.

The builder is not run as part of the CI jobs, as we want them to
run as soon as possible, with the existing images.