curtin: make curtainer use ubuntu-minimal images
Make curtainer use the minimal Ubuntu daily images instead of the full
ones. This speeds up the curtainer run, as the minimal images are
smaller to download and start faster as they ship with less services.

This assumes the ubuntu-minimal-daily remote is correctly setup on the
Jenkins nodes. This is done by running:

  lxc remote add --protocol simplestreams ubuntu-minimal-daily \
      https://cloud-images.ubuntu.com/minimal/daily/

The jenkins-slave-setup Ansible playbook takes care of this, and also
sets up the ubuntu-minimal remote for released images.