git-ubuntu: drop snapcraft packaging distinction
Now that we're switching to a Focal build base, snapcraft is always a
snap so requesting a specific source for snapcraft is moot. In any case,
snapcraft is now installed by the snap.sh wrapper in the git-ubuntu
source tree, rather than being expected to be present in the build
environment.