git-ubuntu: reset VM before test
To ensure that tests are run in a clean environment, reset (ie.
recreate) the VM after the build and before running the tests.

This is particularly important in the new refactoring of the git-ubuntu
snap build to be based on core20. Since it is now based on the deb, and
build dependencies are installed in the VM to build the deb, it's
possible that those build dependencies will be accidentally depended
upon by the snap. In this case, the test of the snap might pass when it
should fail. By resetting the VM, the host system will be cleaner for
testing the snap.